stripe_webhook/requirements.txt
# stripe_webhook/requirements.txt
stripe
functions-framework
google-cloud-firestore